requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.0
httplib2==0.22.0
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
keybert==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.32.0
sentence-transformers==4.1.0
setuptools==80.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0